Decouple controller functions from the database and split up the config files
