+++ backend/.eslintrc.json
@@ -9,6 +9,7 @@
         "sourceType": "module"
     },
     "rules": {
-        "semi": [2, "always"]
+        "semi": [2, "always"],
+        "indent": ["error", 4]
     }
 }
