+++ backend/src/configs/app.js
@@ -1,0 +1,2 @@
+/* Define the port that the web service should listen on */
+export const SERVICE_PORT = process.env.SERVICE_PORT || 3001;
