+++ backend/src/configs/database.js
@@ -1,6 +1,3 @@
-/* Define the port that the web service should listen on */
-export const SERVICE_PORT = process.env.SERVICE_PORT || 3001;
-
 /* Define the database name */
 export const DATABASE_NAME = process.env.DATABASE_NAME || 'todo';
