+++ backend/src/configs/index.js
@@ -1,0 +1,2 @@
+export * from './app.js';
+export * from './database.js';
