+++ backend/src/functions/index.js
@@ -1,0 +1,4 @@
+export { default as createTodo } from './create-todo.js';
+export { default as deleteTodo } from './delete-todo.js';
+export { default as fetchTodos } from './fetch-todos.js';
+export { default as updateTodo } from './update-todo.js';
