+++ backend/src/routes/api.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { index, create, update, destroy } from '../controllers/todo.controller.js';
+import { index, create, update, destroy } from '../controllers/todo-controller.js';
 
 const router = Router();
