+++ backend/tests/todo.test.js
@@ -2,8 +2,8 @@ import request from 'supertest';
 import mongoose from 'mongoose';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import { randomUUID } from 'crypto';
-import { app } from '../src/app.js';
-import { Todo } from '../src/models/todo.model.js';
+import { app } from '../src/bootstrap/app.js';
+import { Todo } from '../src/models/todo.js';
 
 let mongo;
